# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mohir kids Api Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "bakhriddin.inamov@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://mohir-kids.test/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/children/me": {
            "get": {
                "tags": [
                    "Children"
                ],
                "summary": "Children me",
                "description": "Returns children model",
                "operationId": "childrenMe",
                "parameters": [
                    {
                        "name": "bearer",
                        "in": "header",
                        "description": "Children access token",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChildrenModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/parent/login": {
            "post": {
                "tags": [
                    "Parents"
                ],
                "summary": "Login parents",
                "description": "Returns message",
                "operationId": "storeProject",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ParentLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParentResponseObject"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Parent not Found"
                    }
                }
            }
        },
        "/parent/create": {
            "post": {
                "tags": [
                    "Parents"
                ],
                "summary": "Parent create",
                "description": "Returns message",
                "operationId": "parentCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ParentCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParentResponseObject"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/parent/login/attempt": {
            "post": {
                "tags": [
                    "Parents"
                ],
                "summary": "Parent login attempt",
                "description": "Returns message",
                "operationId": "loginAttempt",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckSmsTokenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParentAccessTokenResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/parent/me": {
            "get": {
                "tags": [
                    "Parents"
                ],
                "summary": "Parent me",
                "description": "Returns parent model",
                "operationId": "parentMe",
                "parameters": [
                    {
                        "name": "bearer",
                        "in": "header",
                        "description": "Access token",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParentModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/parent/childrens": {
            "get": {
                "tags": [
                    "Parents"
                ],
                "summary": "Parent childrens",
                "description": "Returns parent childrens",
                "operationId": "parentChildrens",
                "parameters": [
                    {
                        "name": "bearer",
                        "in": "header",
                        "description": "Parent access token",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChildrenModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CheckSmsTokenRequest": {
                "title": "Check sms token request",
                "description": "Check sms token request body",
                "required": [
                    "phone",
                    "code"
                ],
                "properties": {
                    "phone": {
                        "title": "phone",
                        "description": "Parent phone number",
                        "type": "string",
                        "example": "+998950330455"
                    },
                    "code": {
                        "title": "code",
                        "description": "Code from sms",
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
...
            "ChildrenModel": {
                "title": "Children",
                "description": "Children model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int32",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "format": "string",
                        "example": "Nematjon"
                    },
                    "birthdate": {
                        "title": "birthdate",
                        "description": "birthdate",
                        "type": "string",
                        "format": "date",
                        "example": "2022-01-01"
                    },
                    "grade": {
                        "title": "grade",
                        "description": "grade",
                        "type": "string",
                        "format": "string",
                        "example": "7-sinf"
                    },
                    "whois": {
                        "title": "whois",
                        "description": "whois",
                        "type": "string",
                        "format": "string",
                        "example": "Dadasi"
                    },
                    "image": {
                        "title": "image",
                        "description": "image url",
                        "type": "string",
                        "format": "string",
                        "example": "https://mohir-kids.uz/blank.png"
                    },
                    "code": {
                        "title": "code",
                        "description": "code",
                        "type": "integer",
                        "format": "int64",
                        "example": "abrorakmalov#1275"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "integer",
                        "format": "int64",
                        "example": "123456"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ChildrenModel"
                }
            },
...
            "ParentModel": {
                "title": "Parent",
                "description": "Parent model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "Parent name",
                        "type": "string",
                        "format": "string",
                        "example": "Nematjon"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "Parent phone number",
                        "type": "string",
                        "format": "string",
                        "example": "+998950330455"
                    },
                    "image": {
                        "title": "image",
                        "description": "Image url",
                        "type": "string",
                        "format": "string",
                        "example": "https://upload.wikimedia.org/wikipedia/commons/7/70/User_icon_BLACK-01.png"
                    },
                    "is_free": {
                        "title": "is_free",
                        "description": "Free user",
                        "type": "boolean",
                        "format": "boolean",
                        "example": "true"
                    },
                    "is_active": {
                        "title": "is_active",
                        "description": "Active user",
                        "type": "boolean",
                        "format": "boolean",
                        "example": "true"
                    },
                    "remember_token": {
                        "title": "remember_token",
                        "description": "Remember token",
                        "type": "string",
                        "format": "string",
                        "example": "asdasdasfasfasd",
                        "nullable": true
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "Created at",
                        "type": "string",
                        "format": "string",
                        "example": "2022-04-04T10:10:10",
                        "nullable": true
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "string",
                        "example": "2022-04-04T10:10:10",
                        "nullable": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ParentModel"
                }
            },
...
            "ParentAccessTokenResponse": {
                "title": "Parent access token response",
                "description": "Parent access token response",
                "required": [
                    "name",
                    "phone"
                ],
                "properties": {
                    "access_token": {
                        "title": "access token",
                        "description": "Access token",
                        "type": "string",
                        "example": ""
                    },
                    "token_type": {
                        "title": "token type",
                        "description": "Access token type",
                        "type": "string",
                        "example": "bearer"
                    },
                    "expires_in": {
                        "title": "expires time",
                        "description": "Expires access token time",
                        "type": "string",
                        "example": "86400"
                    }
                },
                "type": "object"
            },
            "ParentCreateRequest": {
                "title": "Parent create request",
                "description": "Parent create request body",
                "required": [
                    "name",
                    "phone"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Parent name",
                        "type": "string",
                        "example": "Nematjon"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "Parent phone number",
                        "type": "string",
                        "example": "+998950330455"
                    }
                },
                "type": "object"
            },
            "ParentLoginRequest": {
                "title": "Parent login request",
                "description": "Parent login request body",
                "required": [
                    "phone"
                ],
                "properties": {
                    "phone": {
                        "title": "phone",
                        "description": "Parent phone number",
                        "type": "string",
                        "example": "+998950330455"
                    }
                },
                "type": "object"
            },
            "ParentResponseObject": {
                "title": "ParentResponse",
                "description": "Parent response",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status code",
                        "type": "string",
                        "example": "201"
                    },
                    "message": {
                        "title": "message",
                        "description": "Message",
                        "type": "string",
                        "example": "Sms successfully sent"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ParentResponseObject"
                }
            },
...
        }
    }
}